xxhsum: split sanity check
It now exists in xsum_sanity_check.c

Also add `XSUM_NO_TESTS` option, which instead prints "This version of xxhsum
is not verified." to stderr whenever XSUM_sanityCheck() is called.
